Cleanly stop a service group if a service exits

+++ modules/services.nix
@@ -73,11 +73,13 @@ let
             pid=$!
             echo $pid > "$PRJ_DATA_DIR/pids/${gName}.pid"
             on_stop() {
-                kill -TERM $pid
+                if ps -p $pid > /dev/null; then
+                  kill -TERM $pid
+                fi
                 rm "$PRJ_DATA_DIR/pids/${gName}.pid"
                 wait $pid
             }
-            trap "on_stop" SIGINT SIGTERM SIGHUP
+            trap "on_stop" SIGINT SIGTERM SIGHUP EXIT
             wait $pid
           '').outPath;
       }
